Move chat runtime lifecycle files
